Finished ComputerDAOImpl, and refactored IDs ; Integer are now Long.

+++ src/com/excilys/formation/cdb/mapper/CompanyMapper.java
@@ -13,7 +13,7 @@ public class CompanyMapper {
         Company c = new Company();
         try {
             while (rs.next()) {
-                c.setId(rs.getInt("id"));
+                c.setId(rs.getLong("id"));
                 c.setName(rs.getString("name"));
             }
         }
@@ -28,7 +28,7 @@ public class CompanyMapper {
         try {
             while (rs.next()) {
                 Company c = new Company();
-                c.setId(rs.getInt("id"));
+                c.setId(rs.getLong("id"));
                 c.setName(rs.getString("name"));
                 companies.add(c);
             }
