+++ src/com/excilys/formation/cdb/persistance/dao/CompanyDAO.java
@@ -6,9 +6,9 @@ import com.excilys.formation.cdb.model.Company;
 
 public interface CompanyDAO {
 	
-	public abstract int getNumberOfCompanies();
+	public abstract Long getNumberOfCompanies();
 	
-	public abstract Company getCompany(int id);
+	public abstract Company getCompany(Long id);
 	public abstract List<Company> getCompany(String name, int index, int offset);
 	public abstract List<Company> getCompanies(int index, int stop);
